WIP: Trying to use svc::mk()
