+++ src/app/mod.rs
@@ -14,6 +14,7 @@ mod main;
 mod metric_labels;
 mod outbound;
 mod profiles;
+mod tap;
 
 pub use self::main::Main;
 use addr::{self, Addr};
