+++ src/lib.rs
@@ -1,4 +1,4 @@
-#![deny(warnings)]
+// #![deny(warnings)]
 #![recursion_limit = "128"]
 
 extern crate bytes;
