+++ src/proxy/grpc/mod.rs
@@ -2,4 +2,4 @@ mod body;
 mod service;
 
 pub use self::body::GrpcBody;
-pub use self::service::{req_body_as_payload, req_box_body, res_body_as_payload, unauthorized};
+pub use self::service::{req_body_as_payload, req_box_body, res_body_as_payload};
